requirements.txt
Django
django-simple-sso
django-jet
